+++ Source/ACE.Common/GameConfiguration.cs
@@ -17,6 +17,10 @@ namespace ACE.Common
 
         [System.ComponentModel.DefaultValue(false)]
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Populate)]
+        public bool ServerPerformanceMonitorAutoStart { get; set; }
+
+        [System.ComponentModel.DefaultValue(false)]
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Populate)]
         public bool WorldDatabasePrecaching { get; set; }
 
         [System.ComponentModel.DefaultValue(true)]