+++ changelog.md
@@ -1,5 +1,9 @@
 # ACEmulator Change Log
 
+### 2019-04-03
+[Mag-nus]
+* /serverperformance command added. Optional parameters: start, stop, reset
+
 ### 2019-04-02
 [Ripley]
 * Change the way GiveObjecttoNPC and HandleNPCReceiveItem deal with emotes.